+++ Ids.cs
@@ -11,7 +11,8 @@ internal static class Ids
     internal const string MaxClockSpeed = "cpu.maxclock";
 
     internal const string TotalUsage = "cpu.totalusage";
-    internal const string TotalClock = "cpu.totalclock";
+    internal const string TotalTemperature = "cpu.totaltemperature";
+    internal const string TotalPower = "cpu.totalPower";
   }
 
   internal static class Gpu
@@ -22,8 +23,10 @@ internal static class Ids
     internal const string RefreshRate = "gpu.refreshrate";
     internal const string HorizontalResolution = "gpu.hres";
     internal const string VerticalResolution = "gpu.vres";
-    internal const string Usage3D = "gpu.usage.3d";
-    internal const string UsedMemory = "gpu.memory.used";
+    internal const string UsageCore = "gpu.usage.core";
+    internal const string UsageMemory = "gpu.usage.memory";
+    internal const string Power = "gpu.power";
+    internal const string Temperature = "gpu.temperature";
   }
 
   internal static class Memory
@@ -55,7 +58,6 @@ internal static class Ids
   internal static class Groups
   {
     internal const string CpuGroupIndividual = "group.cpu";
-    internal const string CpuGroupOverall = "group.cpu.overall";
     internal const string GpuGroupIndividual = "group.gpu";
     internal const string MemoryGroupIndividual = "group.memory";
     internal const string MemoryGroupOverall = "group.memory.overall";