+++ Plugin.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using LibreHardwareMonitor.Hardware;
 using MoBro.Plugin.MoBroHardwareMonitor.DataCollectors;
 using MoBro.Plugin.MoBroHardwareMonitor.Model;
 using MoBro.Plugin.SDK;
@@ -8,15 +9,16 @@ using MoBro.Plugin.SDK.Services;
 
 namespace MoBro.Plugin.MoBroHardwareMonitor;
 
-public sealed class MoBroHardwareMonitor : IMoBroPlugin
+public sealed class Plugin : IMoBroPlugin, IDisposable
 {
   private static readonly TimeSpan InitialDelay = TimeSpan.FromSeconds(2);
   private const int DefaultUpdateFrequencyMs = 1000;
 
-  private readonly IHardwareInfoCollector _hardwareInfoCollector;
-  private readonly IHardwareMonitor _hardwareMonitor;
   private readonly IMoBroService _service;
   private readonly IMoBroScheduler _scheduler;
+  private IHardwareInfoCollector _hardwareInfoCollector;
+  private IHardwareMonitor _hardwareMonitor;
+  private Computer _computer;
 
   private readonly int _updateFrequency;
   private readonly bool _monitorCpu;
@@ -25,12 +27,10 @@ public sealed class MoBroHardwareMonitor : IMoBroPlugin
   private readonly int _numProcesses;
   private readonly string _processesSort;
 
-  public MoBroHardwareMonitor(IMoBroService service, IMoBroSettings settings, IMoBroScheduler scheduler)
+  public Plugin(IMoBroService service, IMoBroSettings settings, IMoBroScheduler scheduler)
   {
     _service = service;
     _scheduler = scheduler;
-    _hardwareInfoCollector = new HardwareInfoCollector();
-    _hardwareMonitor = new HardwareMonitor();
 
     _updateFrequency = settings.GetValue("update_frequency", DefaultUpdateFrequencyMs);
     _monitorCpu = settings.GetValue<bool>("cpu_metrics", true);
@@ -42,6 +42,24 @@ public sealed class MoBroHardwareMonitor : IMoBroPlugin
 
   public void Init()
   {
+    // init LibreHardwareMonitor
+    _computer = new Computer
+    {
+      IsCpuEnabled = _monitorCpu,
+      IsGpuEnabled = _monitorGpu,
+      IsMemoryEnabled = false,
+      IsMotherboardEnabled = false,
+      IsStorageEnabled = false,
+      IsNetworkEnabled = false,
+      IsControllerEnabled = false,
+      IsPsuEnabled = false,
+      IsBatteryEnabled = false
+    };
+    _computer.Open();
+
+    _hardwareInfoCollector = new HardwareInfoCollector();
+    _hardwareMonitor = new HardwareMonitor(_computer);
+
     // static metrics
     Register(_hardwareInfoCollector.GetSystem());
     Register(_hardwareInfoCollector.GetProcessors());
@@ -50,7 +68,7 @@ public sealed class MoBroHardwareMonitor : IMoBroPlugin
 
     // dynamic metrics
     Register(_hardwareMonitor.GetSystem());
-    if (_monitorCpu) Register(_hardwareMonitor.GetProcessor());
+    if (_monitorCpu) Register(_hardwareMonitor.GetProcessors());
     if (_monitorGpu) Register(_hardwareMonitor.GetGraphics());
     if (_monitorRam) Register(_hardwareMonitor.GetMemory());
     if (_numProcesses > 0) Register(_hardwareMonitor.GetProcessesStats(_numProcesses, _processesSort));
@@ -63,7 +81,7 @@ public sealed class MoBroHardwareMonitor : IMoBroPlugin
   {
     _service.UpdateMetricValues(_hardwareMonitor.GetSystem().ToMetricValues());
 
-    if (_monitorCpu) _service.UpdateMetricValues(_hardwareMonitor.GetProcessor().ToMetricValues());
+    if (_monitorCpu) _service.UpdateMetricValues(_hardwareMonitor.GetProcessors().SelectMany(c => c.ToMetricValues()));
     if (_monitorGpu) _service.UpdateMetricValues(_hardwareMonitor.GetGraphics().SelectMany(g => g.ToMetricValues()));
     if (_monitorRam) _service.UpdateMetricValues(_hardwareMonitor.GetMemory().ToMetricValues());
     if (_numProcesses > 0) _service.UpdateMetricValues(_hardwareMonitor.GetProcessesStats(_numProcesses, _processesSort) .SelectMany(g => g.ToMetricValues()));
@@ -90,5 +108,10 @@ public sealed class MoBroHardwareMonitor : IMoBroPlugin
     {
       _service.UpdateMetricValue(metricValue);
     }
+  }
+
+  public void Dispose()
+  {
+    _computer.Close();
   }
 }