+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Included LibreHardwareMonitor Lib
+- Return metrics for each individual CPU
+- New metrics: GPU temperature, GPU core, GPU, power, GPU memory load, CPU temperature, CPU power
+
+### Changed
+
+- Removed metrics: CPU clock
+
 ## 0.3.0 - 2024-03-15
 
 ### Changed